Update javascript tests for new data in portal
