--- pdp/static/js/__mocks__/data-services/vic_gen1_app.js
@@ -1,1628 +1,0 @@
-// Mock data-services module for testing gridded_observations_app
-// For tests only, so is a pure Node module
-// Note this does not follow the arrangement for
-// [Jest module "manual mocks"] (https://jestjs.io/docs/en/manual-mocks)
-
-var convert = require('xml-js');
-
-var mockHelpers = require('../../__test__/mock-helpers');
-var makeMockGet = mockHelpers.makeMockGet;
-
-// Mocking presently only for wind_day_TPS_NWNA_v1_historical_19450101-20121231
-
-var catalog = {
-    "5var_day_CCSM3_A1B_run1_19500101-20991231": "https://data.fake.org/data/hydro_model_out/5var_day_CCSM3_A1B_run1_19500101-20991231.nc"
-};
-var getCatalog = makeMockGet('Catalog', catalog);
-
-
-var metadata = {
-    "units": "m", "max": 1.12804, "min": 0.0111
-};
-var getMetadata = makeMockGet('Metadata', metadata);
-
-
-var rasterAccordionMenuData = {
-    "A1B": {
-        "CCSM3": {
-            "pr": "pr-tasmin-tasmax-wind_day_CCSM3_A1B_run1_19500101-21001231",
-            "bf": "5var_day_CCSM3_A1B_run1_19500101-20991231",
-            "aet": "5var_day_CCSM3_A1B_run1_19500101-20991231",
-            "sm": "5var_day_CCSM3_A1B_run1_19500101-20991231",
-            "swe": "5var_day_CCSM3_A1B_run1_19500101-20991231",
-            "R": "5var_day_CCSM3_A1B_run1_19500101-20991231",
-            "tasmin": "pr-tasmin-tasmax-wind_day_CCSM3_A1B_run1_19500101-21001231",
-            "tasmax": "pr-tasmin-tasmax-wind_day_CCSM3_A1B_run1_19500101-21001231",
-            "wind": "pr-tasmin-tasmax-wind_day_CCSM3_A1B_run1_19500101-21001231"
-        },
-    },
-};
-var getRasterAccordionMenuData = makeMockGet(
-    'RasterAccordionMenuData', rasterAccordionMenuData, true);
-
-
-// TODO: Feckin ell, can we cut this down?
-// Maybe just to what the receiver cares about?
-var ncwmsLayerCapablilitiesXml = [
-    '<!DOCTYPE WMT_MS_Capabilities SYSTEM "http://schemas.opengis.net/wms/1.1.1/capabilities_1_1_1.dtd">',
-    '<WMT_MS_Capabilities',
-    '        version="1.1.1"',
-    '        updateSequence="2019-06-13T21:22:17.193Z"',
-    '        xmlns:xlink="http://www.w3.org/1999/xlink">',
-    '    <!-- Service Metadata -->',
-    '    <Service>',
-    '        <!-- The WMT-defined name for this type of service -->',
-    '        <Name>OGC:WMS</Name>',
-    '        <!-- Human-readable title for pick lists -->',
-    '        <Title>My ncWMS server</Title>',
-    '        <!-- Narrative description providing additional information -->',
-    '        <Abstract> </Abstract>',
-    '        <KeywordList>',
-    '            ',
-    '            ',
-    '            <Keyword> </Keyword>',
-    '            ',
-    '        </KeywordList>',
-    '        <!-- Top-level web address of service or service provider. See also OnlineResource',
-    '        elements under <DCPType>. -->',
-    '        <OnlineResource xlink:type="simple" xlink:href=" "/>',
-    '        <!-- Contact information -->',
-    '        <ContactInformation>',
-    '            <ContactPersonPrimary>',
-    '                <ContactPerson> </ContactPerson>',
-    '                <ContactOrganization> </ContactOrganization>',
-    '            </ContactPersonPrimary>',
-    '            <ContactVoiceTelephone> </ContactVoiceTelephone>',
-    '            <ContactElectronicMailAddress> </ContactElectronicMailAddress>',
-    '        </ContactInformation>',
-    '        <!-- Fees or access constraints imposed. -->',
-    '        <Fees>none</Fees>',
-    '        <AccessConstraints>none</AccessConstraints>',
-    '    </Service>',
-    '    <Capability>',
-    '        <Request>',
-    '            <GetCapabilities>',
-    '                <Format>application/vnd.ogc.wms_xml</Format>',
-    '                <DCPType>',
-    '                    <HTTP>',
-    '                        <Get>',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms" />',
-    '                        </Get>',
-    '                    </HTTP>',
-    '                </DCPType>',
-    '            </GetCapabilities>',
-    '            <GetMap>',
-    '                ',
-    '                <Format>image/png</Format>',
-    '                ',
-    '                <Format>image/png;mode=32bit</Format>',
-    '                ',
-    '                <Format>image/gif</Format>',
-    '                ',
-    '                <Format>image/jpeg</Format>',
-    '                ',
-    '                <Format>application/vnd.google-earth.kmz</Format>',
-    '                ',
-    '                <DCPType>',
-    '                    <HTTP>',
-    '                        <Get>',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms" />',
-    '                        </Get>',
-    '                    </HTTP>',
-    '                </DCPType>',
-    '            </GetMap>',
-    '            <GetFeatureInfo>',
-    '                ',
-    '                <Format>image/png</Format>',
-    '                ',
-    '                <Format>text/xml</Format>',
-    '                ',
-    '                <DCPType>',
-    '                    <HTTP>',
-    '                        <Get>',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms" />',
-    '                        </Get>',
-    '                    </HTTP>',
-    '                </DCPType>',
-    '            </GetFeatureInfo>',
-    '        </Request>',
-    '        <Exception>',
-    '            <Format>application/vnd.ogc.se_xml</Format>',
-    '            <!--<Format>application/vnd.ogc.se_inimage</Format>',
-    '            <Format>application/vnd.ogc.se_blank</Format>-->',
-    '        </Exception>',
-    '        ',
-    '        <Layer>',
-    '            <Title>My ncWMS server</Title>',
-    '            ',
-    '            <SRS>EPSG:4326</SRS>',
-    '            ',
-    '            <SRS>CRS:84</SRS>',
-    '            ',
-    '            <SRS>EPSG:41001</SRS>',
-    '            ',
-    '            <SRS>EPSG:27700</SRS>',
-    '            ',
-    '            <SRS>EPSG:3408</SRS>',
-    '            ',
-    '            <SRS>EPSG:3409</SRS>',
-    '            ',
-    '            <SRS>EPSG:3857</SRS>',
-    '            ',
-    '            <SRS>EPSG:32661</SRS>',
-    '            ',
-    '            <SRS>EPSG:32761</SRS>',
-    '            ',
-    '            ',
-    '            ',
-    '            <Layer>',
-    '                <Title>5var_day_CCSM3_A1B_run1_19500101-20991231</Title>',
-    '                ',
-    '                <Layer queryable="1">',
-    '                    <Name>5var_day_CCSM3_A1B_run1_19500101-20991231/sm</Name>',
-    '                    <Title>lwe_thickness_of_moisture_content_of_soil_layer</Title>',
-    '                    <Abstract>Total Soil Moisture (3 layers)</Abstract>',
-    '                    ',
-    '                    <LatLonBoundingBox minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    <BoundingBox SRS="EPSG:4326" minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    ',
-    '                    <Dimension name="time" units="ISO8601"/>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Extent name="time" multipleValues="1" current="1" default="2019-06-13T00:00:00.000Z">',
-    '                        ',
-    '                            ',
-    '                            ',
-    '                                ',
-    '                                1950-01-01T00:00:00.000Z/2099-12-31T00:00:00.000Z/P1D',
-    '                            ',
-    '                        ',
-    '                    </Extent>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg</Name>',
-    '                        <Title>boxfill/alg</Title>',
-    '                        <Abstract>boxfill style, using the alg palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=alg"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/greyscale</Name>',
-    '                        <Title>boxfill/greyscale</Title>',
-    '                        <Abstract>boxfill style, using the greyscale palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=greyscale"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ncview</Name>',
-    '                        <Title>boxfill/ncview</Title>',
-    '                        <Abstract>boxfill style, using the ncview palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=ncview"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam</Name>',
-    '                        <Title>boxfill/occam</Title>',
-    '                        <Abstract>boxfill style, using the occam palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=occam"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/yellow_red</Name>',
-    '                        <Title>boxfill/yellow_red</Title>',
-    '                        <Abstract>boxfill style, using the yellow_red palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=yellow_red"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue7_red3</Name>',
-    '                        <Title>boxfill/blue7_red3</Title>',
-    '                        <Abstract>boxfill style, using the blue7_red3 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blue7_red3"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/red_yellow</Name>',
-    '                        <Title>boxfill/red_yellow</Title>',
-    '                        <Abstract>boxfill style, using the red_yellow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=red_yellow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=lightblue_darkblue_log"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_inv</Name>',
-    '                        <Title>boxfill/occam_inv</Title>',
-    '                        <Abstract>boxfill style, using the occam_inv palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=occam_inv"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/sst_36</Name>',
-    '                        <Title>boxfill/sst_36</Title>',
-    '                        <Abstract>boxfill style, using the sst_36 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=sst_36"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ferret</Name>',
-    '                        <Title>boxfill/ferret</Title>',
-    '                        <Abstract>boxfill style, using the ferret palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=ferret"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_pastel-30</Name>',
-    '                        <Title>boxfill/occam_pastel-30</Title>',
-    '                        <Abstract>boxfill style, using the occam_pastel-30 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=occam_pastel-30"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/redblue</Name>',
-    '                        <Title>boxfill/redblue</Title>',
-    '                        <Abstract>boxfill style, using the redblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=redblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_green</Name>',
-    '                        <Title>boxfill/brown_green</Title>',
-    '                        <Abstract>boxfill style, using the brown_green palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=brown_green"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blueheat</Name>',
-    '                        <Title>boxfill/blueheat</Title>',
-    '                        <Abstract>boxfill style, using the blueheat palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blueheat"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue4_red6</Name>',
-    '                        <Title>boxfill/blue4_red6</Title>',
-    '                        <Abstract>boxfill style, using the blue4_red6 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blue4_red6"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg2</Name>',
-    '                        <Title>boxfill/alg2</Title>',
-    '                        <Abstract>boxfill style, using the alg2 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=alg2"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log_nc</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log_nc</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log_nc palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=lightblue_darkblue_log_nc"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_blue</Name>',
-    '                        <Title>boxfill/brown_blue</Title>',
-    '                        <Abstract>boxfill style, using the brown_blue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=brown_blue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue6_red4</Name>',
-    '                        <Title>boxfill/blue6_red4</Title>',
-    '                        <Abstract>boxfill style, using the blue6_red4 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blue6_red4"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_brown</Name>',
-    '                        <Title>boxfill/blue_brown</Title>',
-    '                        <Abstract>boxfill style, using the blue_brown palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blue_brown"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_darkred</Name>',
-    '                        <Title>boxfill/blue_darkred</Title>',
-    '                        <Abstract>boxfill style, using the blue_darkred palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=blue_darkred"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue</Name>',
-    '                        <Title>boxfill/lightblue_darkblue</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=lightblue_darkblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/rainbow</Name>',
-    '                        <Title>boxfill/rainbow</Title>',
-    '                        <Abstract>boxfill style, using the rainbow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/sm&amp;PALETTE=rainbow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    ',
-    '                </Layer>',
-    '                ',
-    '                <Layer queryable="1">',
-    '                    <Name>5var_day_CCSM3_A1B_run1_19500101-20991231/aet</Name>',
-    '                    <Title>lwe_thickness_of_water_evaporation_amount</Title>',
-    '                    <Abstract>Actual Evapotranspiration</Abstract>',
-    '                    ',
-    '                    <LatLonBoundingBox minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    <BoundingBox SRS="EPSG:4326" minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    ',
-    '                    <Dimension name="time" units="ISO8601"/>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Extent name="time" multipleValues="1" current="1" default="2019-06-13T00:00:00.000Z">',
-    '                        ',
-    '                            ',
-    '                            ',
-    '                                ',
-    '                                1950-01-01T00:00:00.000Z/2099-12-31T00:00:00.000Z/P1D',
-    '                            ',
-    '                        ',
-    '                    </Extent>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg</Name>',
-    '                        <Title>boxfill/alg</Title>',
-    '                        <Abstract>boxfill style, using the alg palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=alg"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/greyscale</Name>',
-    '                        <Title>boxfill/greyscale</Title>',
-    '                        <Abstract>boxfill style, using the greyscale palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=greyscale"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ncview</Name>',
-    '                        <Title>boxfill/ncview</Title>',
-    '                        <Abstract>boxfill style, using the ncview palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=ncview"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam</Name>',
-    '                        <Title>boxfill/occam</Title>',
-    '                        <Abstract>boxfill style, using the occam palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=occam"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/yellow_red</Name>',
-    '                        <Title>boxfill/yellow_red</Title>',
-    '                        <Abstract>boxfill style, using the yellow_red palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=yellow_red"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue7_red3</Name>',
-    '                        <Title>boxfill/blue7_red3</Title>',
-    '                        <Abstract>boxfill style, using the blue7_red3 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blue7_red3"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/red_yellow</Name>',
-    '                        <Title>boxfill/red_yellow</Title>',
-    '                        <Abstract>boxfill style, using the red_yellow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=red_yellow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=lightblue_darkblue_log"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_inv</Name>',
-    '                        <Title>boxfill/occam_inv</Title>',
-    '                        <Abstract>boxfill style, using the occam_inv palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=occam_inv"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/sst_36</Name>',
-    '                        <Title>boxfill/sst_36</Title>',
-    '                        <Abstract>boxfill style, using the sst_36 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=sst_36"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ferret</Name>',
-    '                        <Title>boxfill/ferret</Title>',
-    '                        <Abstract>boxfill style, using the ferret palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=ferret"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_pastel-30</Name>',
-    '                        <Title>boxfill/occam_pastel-30</Title>',
-    '                        <Abstract>boxfill style, using the occam_pastel-30 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=occam_pastel-30"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/redblue</Name>',
-    '                        <Title>boxfill/redblue</Title>',
-    '                        <Abstract>boxfill style, using the redblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=redblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_green</Name>',
-    '                        <Title>boxfill/brown_green</Title>',
-    '                        <Abstract>boxfill style, using the brown_green palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=brown_green"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blueheat</Name>',
-    '                        <Title>boxfill/blueheat</Title>',
-    '                        <Abstract>boxfill style, using the blueheat palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blueheat"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue4_red6</Name>',
-    '                        <Title>boxfill/blue4_red6</Title>',
-    '                        <Abstract>boxfill style, using the blue4_red6 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blue4_red6"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg2</Name>',
-    '                        <Title>boxfill/alg2</Title>',
-    '                        <Abstract>boxfill style, using the alg2 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=alg2"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log_nc</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log_nc</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log_nc palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=lightblue_darkblue_log_nc"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_blue</Name>',
-    '                        <Title>boxfill/brown_blue</Title>',
-    '                        <Abstract>boxfill style, using the brown_blue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=brown_blue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue6_red4</Name>',
-    '                        <Title>boxfill/blue6_red4</Title>',
-    '                        <Abstract>boxfill style, using the blue6_red4 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blue6_red4"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_brown</Name>',
-    '                        <Title>boxfill/blue_brown</Title>',
-    '                        <Abstract>boxfill style, using the blue_brown palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blue_brown"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_darkred</Name>',
-    '                        <Title>boxfill/blue_darkred</Title>',
-    '                        <Abstract>boxfill style, using the blue_darkred palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=blue_darkred"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue</Name>',
-    '                        <Title>boxfill/lightblue_darkblue</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=lightblue_darkblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/rainbow</Name>',
-    '                        <Title>boxfill/rainbow</Title>',
-    '                        <Abstract>boxfill style, using the rainbow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/aet&amp;PALETTE=rainbow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    ',
-    '                </Layer>',
-    '                ',
-    '                <Layer queryable="1">',
-    '                    <Name>5var_day_CCSM3_A1B_run1_19500101-20991231/R</Name>',
-    '                    <Title>lwe_thickness_of_surface_runoff_amount</Title>',
-    '                    <Abstract>Runoff</Abstract>',
-    '                    ',
-    '                    <LatLonBoundingBox minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    <BoundingBox SRS="EPSG:4326" minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    ',
-    '                    <Dimension name="time" units="ISO8601"/>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Extent name="time" multipleValues="1" current="1" default="2019-06-13T00:00:00.000Z">',
-    '                        ',
-    '                            ',
-    '                            ',
-    '                                ',
-    '                                1950-01-01T00:00:00.000Z/2099-12-31T00:00:00.000Z/P1D',
-    '                            ',
-    '                        ',
-    '                    </Extent>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg</Name>',
-    '                        <Title>boxfill/alg</Title>',
-    '                        <Abstract>boxfill style, using the alg palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=alg"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/greyscale</Name>',
-    '                        <Title>boxfill/greyscale</Title>',
-    '                        <Abstract>boxfill style, using the greyscale palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=greyscale"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ncview</Name>',
-    '                        <Title>boxfill/ncview</Title>',
-    '                        <Abstract>boxfill style, using the ncview palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=ncview"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam</Name>',
-    '                        <Title>boxfill/occam</Title>',
-    '                        <Abstract>boxfill style, using the occam palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=occam"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/yellow_red</Name>',
-    '                        <Title>boxfill/yellow_red</Title>',
-    '                        <Abstract>boxfill style, using the yellow_red palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=yellow_red"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue7_red3</Name>',
-    '                        <Title>boxfill/blue7_red3</Title>',
-    '                        <Abstract>boxfill style, using the blue7_red3 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blue7_red3"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/red_yellow</Name>',
-    '                        <Title>boxfill/red_yellow</Title>',
-    '                        <Abstract>boxfill style, using the red_yellow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=red_yellow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=lightblue_darkblue_log"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_inv</Name>',
-    '                        <Title>boxfill/occam_inv</Title>',
-    '                        <Abstract>boxfill style, using the occam_inv palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=occam_inv"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/sst_36</Name>',
-    '                        <Title>boxfill/sst_36</Title>',
-    '                        <Abstract>boxfill style, using the sst_36 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=sst_36"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ferret</Name>',
-    '                        <Title>boxfill/ferret</Title>',
-    '                        <Abstract>boxfill style, using the ferret palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=ferret"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_pastel-30</Name>',
-    '                        <Title>boxfill/occam_pastel-30</Title>',
-    '                        <Abstract>boxfill style, using the occam_pastel-30 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=occam_pastel-30"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/redblue</Name>',
-    '                        <Title>boxfill/redblue</Title>',
-    '                        <Abstract>boxfill style, using the redblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=redblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_green</Name>',
-    '                        <Title>boxfill/brown_green</Title>',
-    '                        <Abstract>boxfill style, using the brown_green palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=brown_green"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blueheat</Name>',
-    '                        <Title>boxfill/blueheat</Title>',
-    '                        <Abstract>boxfill style, using the blueheat palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blueheat"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue4_red6</Name>',
-    '                        <Title>boxfill/blue4_red6</Title>',
-    '                        <Abstract>boxfill style, using the blue4_red6 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blue4_red6"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg2</Name>',
-    '                        <Title>boxfill/alg2</Title>',
-    '                        <Abstract>boxfill style, using the alg2 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=alg2"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log_nc</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log_nc</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log_nc palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=lightblue_darkblue_log_nc"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_blue</Name>',
-    '                        <Title>boxfill/brown_blue</Title>',
-    '                        <Abstract>boxfill style, using the brown_blue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=brown_blue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue6_red4</Name>',
-    '                        <Title>boxfill/blue6_red4</Title>',
-    '                        <Abstract>boxfill style, using the blue6_red4 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blue6_red4"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_brown</Name>',
-    '                        <Title>boxfill/blue_brown</Title>',
-    '                        <Abstract>boxfill style, using the blue_brown palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blue_brown"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_darkred</Name>',
-    '                        <Title>boxfill/blue_darkred</Title>',
-    '                        <Abstract>boxfill style, using the blue_darkred palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=blue_darkred"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue</Name>',
-    '                        <Title>boxfill/lightblue_darkblue</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=lightblue_darkblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/rainbow</Name>',
-    '                        <Title>boxfill/rainbow</Title>',
-    '                        <Abstract>boxfill style, using the rainbow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/R&amp;PALETTE=rainbow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    ',
-    '                </Layer>',
-    '                ',
-    '                <Layer queryable="1">',
-    '                    <Name>5var_day_CCSM3_A1B_run1_19500101-20991231/bf</Name>',
-    '                    <Title>lwe_thickness_of_subsurface_runoff_amount</Title>',
-    '                    <Abstract>Base Flow</Abstract>',
-    '                    ',
-    '                    <LatLonBoundingBox minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    <BoundingBox SRS="EPSG:4326" minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    ',
-    '                    <Dimension name="time" units="ISO8601"/>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Extent name="time" multipleValues="1" current="1" default="2019-06-13T00:00:00.000Z">',
-    '                        ',
-    '                            ',
-    '                            ',
-    '                                ',
-    '                                1950-01-01T00:00:00.000Z/2099-12-31T00:00:00.000Z/P1D',
-    '                            ',
-    '                        ',
-    '                    </Extent>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg</Name>',
-    '                        <Title>boxfill/alg</Title>',
-    '                        <Abstract>boxfill style, using the alg palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=alg"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/greyscale</Name>',
-    '                        <Title>boxfill/greyscale</Title>',
-    '                        <Abstract>boxfill style, using the greyscale palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=greyscale"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ncview</Name>',
-    '                        <Title>boxfill/ncview</Title>',
-    '                        <Abstract>boxfill style, using the ncview palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=ncview"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam</Name>',
-    '                        <Title>boxfill/occam</Title>',
-    '                        <Abstract>boxfill style, using the occam palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=occam"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/yellow_red</Name>',
-    '                        <Title>boxfill/yellow_red</Title>',
-    '                        <Abstract>boxfill style, using the yellow_red palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=yellow_red"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue7_red3</Name>',
-    '                        <Title>boxfill/blue7_red3</Title>',
-    '                        <Abstract>boxfill style, using the blue7_red3 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blue7_red3"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/red_yellow</Name>',
-    '                        <Title>boxfill/red_yellow</Title>',
-    '                        <Abstract>boxfill style, using the red_yellow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=red_yellow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=lightblue_darkblue_log"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_inv</Name>',
-    '                        <Title>boxfill/occam_inv</Title>',
-    '                        <Abstract>boxfill style, using the occam_inv palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=occam_inv"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/sst_36</Name>',
-    '                        <Title>boxfill/sst_36</Title>',
-    '                        <Abstract>boxfill style, using the sst_36 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=sst_36"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ferret</Name>',
-    '                        <Title>boxfill/ferret</Title>',
-    '                        <Abstract>boxfill style, using the ferret palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=ferret"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_pastel-30</Name>',
-    '                        <Title>boxfill/occam_pastel-30</Title>',
-    '                        <Abstract>boxfill style, using the occam_pastel-30 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=occam_pastel-30"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/redblue</Name>',
-    '                        <Title>boxfill/redblue</Title>',
-    '                        <Abstract>boxfill style, using the redblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=redblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_green</Name>',
-    '                        <Title>boxfill/brown_green</Title>',
-    '                        <Abstract>boxfill style, using the brown_green palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=brown_green"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blueheat</Name>',
-    '                        <Title>boxfill/blueheat</Title>',
-    '                        <Abstract>boxfill style, using the blueheat palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blueheat"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue4_red6</Name>',
-    '                        <Title>boxfill/blue4_red6</Title>',
-    '                        <Abstract>boxfill style, using the blue4_red6 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blue4_red6"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg2</Name>',
-    '                        <Title>boxfill/alg2</Title>',
-    '                        <Abstract>boxfill style, using the alg2 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=alg2"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log_nc</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log_nc</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log_nc palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=lightblue_darkblue_log_nc"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_blue</Name>',
-    '                        <Title>boxfill/brown_blue</Title>',
-    '                        <Abstract>boxfill style, using the brown_blue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=brown_blue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue6_red4</Name>',
-    '                        <Title>boxfill/blue6_red4</Title>',
-    '                        <Abstract>boxfill style, using the blue6_red4 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blue6_red4"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_brown</Name>',
-    '                        <Title>boxfill/blue_brown</Title>',
-    '                        <Abstract>boxfill style, using the blue_brown palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blue_brown"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_darkred</Name>',
-    '                        <Title>boxfill/blue_darkred</Title>',
-    '                        <Abstract>boxfill style, using the blue_darkred palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=blue_darkred"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue</Name>',
-    '                        <Title>boxfill/lightblue_darkblue</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=lightblue_darkblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/rainbow</Name>',
-    '                        <Title>boxfill/rainbow</Title>',
-    '                        <Abstract>boxfill style, using the rainbow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/bf&amp;PALETTE=rainbow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    ',
-    '                </Layer>',
-    '                ',
-    '                <Layer queryable="1">',
-    '                    <Name>5var_day_CCSM3_A1B_run1_19500101-20991231/swe</Name>',
-    '                    <Title>lwe_thickness_of_surface_snow_amount</Title>',
-    '                    <Abstract>Snow Water Equivalent</Abstract>',
-    '                    ',
-    '                    <LatLonBoundingBox minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    <BoundingBox SRS="EPSG:4326" minx="-127.8125" maxx="-114.375" miny="47.875" maxy="58.0625"/>',
-    '                    ',
-    '                    <Dimension name="time" units="ISO8601"/>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Extent name="time" multipleValues="1" current="1" default="2019-06-13T00:00:00.000Z">',
-    '                        ',
-    '                            ',
-    '                            ',
-    '                                ',
-    '                                1950-01-01T00:00:00.000Z/2099-12-31T00:00:00.000Z/P1D',
-    '                            ',
-    '                        ',
-    '                    </Extent>',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg</Name>',
-    '                        <Title>boxfill/alg</Title>',
-    '                        <Abstract>boxfill style, using the alg palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=alg"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/greyscale</Name>',
-    '                        <Title>boxfill/greyscale</Title>',
-    '                        <Abstract>boxfill style, using the greyscale palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=greyscale"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ncview</Name>',
-    '                        <Title>boxfill/ncview</Title>',
-    '                        <Abstract>boxfill style, using the ncview palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=ncview"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam</Name>',
-    '                        <Title>boxfill/occam</Title>',
-    '                        <Abstract>boxfill style, using the occam palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=occam"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/yellow_red</Name>',
-    '                        <Title>boxfill/yellow_red</Title>',
-    '                        <Abstract>boxfill style, using the yellow_red palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=yellow_red"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue7_red3</Name>',
-    '                        <Title>boxfill/blue7_red3</Title>',
-    '                        <Abstract>boxfill style, using the blue7_red3 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blue7_red3"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/red_yellow</Name>',
-    '                        <Title>boxfill/red_yellow</Title>',
-    '                        <Abstract>boxfill style, using the red_yellow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=red_yellow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=lightblue_darkblue_log"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_inv</Name>',
-    '                        <Title>boxfill/occam_inv</Title>',
-    '                        <Abstract>boxfill style, using the occam_inv palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=occam_inv"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/sst_36</Name>',
-    '                        <Title>boxfill/sst_36</Title>',
-    '                        <Abstract>boxfill style, using the sst_36 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=sst_36"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/ferret</Name>',
-    '                        <Title>boxfill/ferret</Title>',
-    '                        <Abstract>boxfill style, using the ferret palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=ferret"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/occam_pastel-30</Name>',
-    '                        <Title>boxfill/occam_pastel-30</Title>',
-    '                        <Abstract>boxfill style, using the occam_pastel-30 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=occam_pastel-30"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/redblue</Name>',
-    '                        <Title>boxfill/redblue</Title>',
-    '                        <Abstract>boxfill style, using the redblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=redblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_green</Name>',
-    '                        <Title>boxfill/brown_green</Title>',
-    '                        <Abstract>boxfill style, using the brown_green palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=brown_green"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blueheat</Name>',
-    '                        <Title>boxfill/blueheat</Title>',
-    '                        <Abstract>boxfill style, using the blueheat palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blueheat"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue4_red6</Name>',
-    '                        <Title>boxfill/blue4_red6</Title>',
-    '                        <Abstract>boxfill style, using the blue4_red6 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blue4_red6"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/alg2</Name>',
-    '                        <Title>boxfill/alg2</Title>',
-    '                        <Abstract>boxfill style, using the alg2 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=alg2"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue_log_nc</Name>',
-    '                        <Title>boxfill/lightblue_darkblue_log_nc</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue_log_nc palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=lightblue_darkblue_log_nc"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/brown_blue</Name>',
-    '                        <Title>boxfill/brown_blue</Title>',
-    '                        <Abstract>boxfill style, using the brown_blue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=brown_blue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue6_red4</Name>',
-    '                        <Title>boxfill/blue6_red4</Title>',
-    '                        <Abstract>boxfill style, using the blue6_red4 palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blue6_red4"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_brown</Name>',
-    '                        <Title>boxfill/blue_brown</Title>',
-    '                        <Abstract>boxfill style, using the blue_brown palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blue_brown"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/blue_darkred</Name>',
-    '                        <Title>boxfill/blue_darkred</Title>',
-    '                        <Abstract>boxfill style, using the blue_darkred palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=blue_darkred"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/lightblue_darkblue</Name>',
-    '                        <Title>boxfill/lightblue_darkblue</Title>',
-    '                        <Abstract>boxfill style, using the lightblue_darkblue palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=lightblue_darkblue"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    <Style>',
-    '                      <Name>boxfill/rainbow</Name>',
-    '                        <Title>boxfill/rainbow</Title>',
-    '                        <Abstract>boxfill style, using the rainbow palette </Abstract>',
-    '                        <LegendURL width="110" height="264"> ',
-    '                            <Format>image/png</Format> ',
-    '                            <OnlineResource xlink:type="simple" xlink:href="http://pizza.pcic.uvic.ca:8080/ncWMS-PCIC/wms?REQUEST=GetLegendGraphic&amp;LAYER=5var_day_CCSM3_A1B_run1_19500101-20991231/swe&amp;PALETTE=rainbow"/>',
-    '                        </LegendURL>',
-    '                    </Style>',
-    '                    ',
-    '                    ',
-    '                </Layer>',
-    '                 ',
-    '            </Layer>',
-    '            ',
-    '            ',
-    '        </Layer>',
-    '    </Capability>',
-    '</WMT_MS_Capabilities>',
-].join('\n');
-var ncwmsLayerCapabilities = convert.xml2js(ncwmsLayerCapablilitiesXml);
-var getNCWMSLayerCapabilities = makeMockGet('NCWMSLayerCapabilities', ncwmsLayerCapabilities);
-
-
-var ncwmsLayerDDS = [
-    [
-        'Dataset {',
-        '     Int32 time[time = 54787];',
-        ' } 5var_day_CCSM3_A1B_run1_19500101-20991231%2Enc;',
-    ].join('\n')
-];
-var getNcwmsLayerDDS = makeMockGet('NcwmsLayerDDS', ncwmsLayerDDS);
-
-
-var ncwmsLayerDAS = [
-    [
-        'Attributes {',
-        '    NC_GLOBAL {',
-        '        String input_data "Bias Corrected Spatially Disaggregated (Wood et al. 2004; Salathe 2005) downscaled GCM";',
-        '        String grid_resolution "Square grid, 1/16 degrees_north by 1/16 degrees_east";',
-        '        String project_id "Hydrologic Modelling: Generation 1";',
-        '        Int32 nco_openmp_thread_number 1;',
-        '        String reference "Schnorbus et al. (2011)";',
-        '        String title "Gridded hydrological data for the Peace, Campbell, Fraser and Columbia River Basins (BC)";',
-        '        String spatial_domain "Rectangular domain constrained to the extent of the Peace, Campbell, Fraser and Columbia River Basins";',
-        '        String source "Variable Infiltation Capcity Model, Liang et al. 1994;1996, http://www.hydro.washington.edu/Lettenmaier/Models/VIC/Documentation/";',
-        '        String version "1.0";',
-        '        String contact "Markus Schnorbus, mschnorb@uvic.ca, tel:(250) 853-3502";',
-        '        String experiment_id "generated 2009-10-26";',
-        '        String Conventions "CF-1.5";',
-        '        String institution "Pacific Climate Impacts Consortium (PCIC), Victoria, BC, www.pacificclimate.org";',
-        '        String history "Tue Jan 14 16:01:06 2014: ncpdq -a time,lat,lon 5var_daily_CCSM3_A1B_run1_19500101-20991231.nc.nobackup 5var_daily_CCSM3_A1B_run1_19500101-20991231_timemajor.nc.nobackup',
-        'created 2014-01-10 13:40:36";',
-        '    }',
-        '    time {',
-        '        String long_name "Time";',
-        '        String NAME "time";',
-        '        String units "days since 1950-01-01";',
-        '        String calendar "standard";',
-        '        String CLASS "DIMENSION_SCALE";',
-        '        String axis "T";',
-        '    }',
-        '    lon {',
-        '        String long_name "Longitude";',
-        '        String NAME "lon";',
-        '        String units "degrees_east";',
-        '        Float64 actual_range -127.781, -114.406;',
-        '        String CLASS "DIMENSION_SCALE";',
-        '        String axis "X";',
-        '    }',
-        '    lat {',
-        '        String long_name "Latitude";',
-        '        String NAME "lat";',
-        '        String units "degrees_north";',
-        '        Float64 actual_range 47.9062, 58.0312;',
-        '        String CLASS "DIMENSION_SCALE";',
-        '        String axis "Y";',
-        '    }',
-        '    sm {',
-        '        String units "m";',
-        '        Float32 _FillValue nan;',
-        '        String standard_name "lwe_thickness_of_moisture_content_of_soil_layer";',
-        '        String long_name "total soil moisture of all VIC soil layers";',
-        '    }',
-        '    aet {',
-        '        String units "m";',
-        '        Float32 _FillValue nan;',
-        '        String standard_name "lwe_thickness_of_water_evaporation_amount";',
-        '        String long_name "total actual evaporation of all components";',
-        '    }',
-        '    R {',
-        '        String units "m";',
-        '        Float32 _FillValue nan;',
-        '        String standard_name "lwe_thickness_of_surface_runoff_amount";',
-        '        String long_name "runoff from the ground surface";',
-        '    }',
-        '    bf {',
-        '        String units "m";',
-        '        Float32 _FillValue nan;',
-        '        String standard_name "lwe_thickness_of_subsurface_runoff_amount";',
-        '        String long_name "subsurface runoff which takes place below the level of the water table";',
-        '    }',
-        '    swe {',
-        '        String units "m";',
-        '        Float32 _FillValue nan;',
-        '        String standard_name "lwe_thickness_of_surface_snow_amount";',
-        '        String long_name "total ground snow water equivalent";',
-        '    }',
-        '}',
-    ].join('\n')
-];
-var getNcwmsLayerDAS = makeMockGet('NcwmsLayerDAS', ncwmsLayerDAS);
-
-
-var getStationCount = mockHelpers.unexpectedRequest({
-    name: 'getStationCount', log: true, throw: true,
-});
-
-
-var getRecordLength = mockHelpers.unexpectedRequest({
-    name: 'getRecordLength', log: true, throw: true,
-});
-
-
-module.exports = {
-    getCatalog: getCatalog,
-    getMetadata: getMetadata,
-    getRasterAccordionMenuData: getRasterAccordionMenuData,
-    getNCWMSLayerCapabilities: getNCWMSLayerCapabilities,
-    getNcwmsLayerDDS: getNcwmsLayerDDS,
-    getNcwmsLayerDAS: getNcwmsLayerDAS,
-    getStationCount: getStationCount,
-    getRecordLength: getRecordLength,
-};
-
-
